requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.28
mysql-connector-python==8.3.0
python-dotenv==1.0.1
google-genai==0.1.0
email-validator==2.1.0.post1
requests==2.31.0
reportlab==4.4.7